+++ packages/widget-dashboard/src/components/widget-actions/widget-actions.tsx
@@ -84,7 +84,9 @@ export function WidgetActions( {
 													]
 												}
 												render={
-													<HostLink to={ route.to } />
+													<HostLink
+														path={ route.path }
+													/>
 												}
 											/>
 										) : (